Relocate form submission code into index.js
